Removing 18F text from footer
Since this is the only element in the footer besides the logo, you are
ok without it. Though based on the US Web Standards, since 18F’s logo
is clear (not a hard-to-read seal) it feels redundant with both. If you
add to the footer at any point, maybe consider bringing this text back.