Default to secure cookies
Basic OWASP recommendation.

Why standard Rails comments this out by default is beyond me. If I was braver, I would head to rails / rails right now and PR it.

For now, it should continue to be considered a defense in depth control for 18F across the board, especially where Auth is involved. I wonder if we can flip this on the cloud.gov level / buildpack level in the future?