Ensure winning bidders are prompted that they are winning; ensure that losers are prompted that they are losing.
Adding a failing spec to reproduce the following error:

Currently, the "you are the winning bidder" text is
displayed to *all* users when there is at least one bid.

Previously our tests were not capturinig this error because
they only tested presence or non-presence of the "winning
bidder" text in two scenarios: when the user placed bids
and when there are no bids. The latter scenario
insufficiently covered the cases when the "winning bid"
prompt should be present.

This commit presents another scenario, where there are bids,
but from other users. <strikethrough>Currently this spec fails and will
ensure confidence in the subsequent bug fix.</strikethrough>

Created BidsHelper#user_is_winning_bidder?

This method takes instances of User and Presenter::Auction as args.

It returns true if the given user is the winning bidder of the
given auction and vice versa.

Prompt users who are not winning bidder in auction w/ other bids.

Test for text saying user is *not* the winning bidder.