Overriding the web_design_standards_fonts CSS
The CSS was not being generated to properly reference fonts with digest
strings in their paths. This seems to be a result of using the url
instead of font-path function to reference the fonts. This fix seems to
resolve the problem, but not sure if this is an issue with our
configuration or the gem yet.