Check a user is logged in before You Are Winner
Even if the logged in user is not the winner of the auction, the check
will crash if there is no logged in user. The check could also be fixed to
return false if nobody is logged in, but will just make this change for
now instead since I don't want to modify that method yet.