Create a separate confirmation step
In order to run the confirm bid through the same PlaceBid validation but
not create a record in the database, I have created a new method
`dry_run` which creates a `PlaceBid` object but does not save it to the
database. This seems better than saving a provisional record to the DB
that might not be confirmed.