Run Application using Docker
Add a Docker Compose configuration file to spin up Postgres and the
Rails application in Docker containers.

Update database configuration so it works when the database is on a
separate host, as in the docker compose configuration, or on localhost,
as when Travis CI runs the tests.

The database name, username, password, and host can be set in
environment variables.  The docker compose configuration sets the
environment variables to use a separate postgres container.