A first cut at reorganizing Cucumber steps
The goal is to group certain types of steps together by functionality
rather than the features/scenarios they are used in. So, all steps for
navigating to a page (or checking you are on one) are in the same place
and so on. This will hopefully make it easier to see when a step like
"show the auction start time" is implemented independently in several steps.