Added copy for winners page
This commit adds the copy for the winners page and would close #578.

In this commit, I extracted the FAQ as a partial to make it easier to keep
track of things. And then, because there are no actual FAQs right now, I
commented the render tag. When we add FAQs, all we'll need to do is
uncomment the render tag.