Remove unused merriweather font styles
The issue with keeping them us that we don't actually have these fonts in this
project so the `font_url` rules are referencing files that do not exist.

These were overriding some rules from USWDS that also references these fonts at
a different location where they do not exist.

If we ever want to use these, we will have to override *and* copy over the font files