Change the failing feature step to use set var
For some reason, I think the issue has something to do with chaining two
business time calculations back to back (maybe there is some rounding or
other adjustment after save in the record). So I just add the business
days offset for delivery to the @end_day variable. This seems to work?