Added a model for doing deadline calculations
I wasn't happy to see we were doing the delivery_due_at calculations
separately in both the AuctionParser and the business_controller that is
called by the frontend javascript, especially since the AuctionParser
one seemed to default to the end of the workday. So I created a single
class that is called by both of those for doing the date time
calculations. This lets us unit test it.

While I was testing it, I discovered that the prior version of
business_time we had pinned to was actually wrong with its timezone
handling. So I upgraded to the latest version and all tests seem to pass now.