Fix the holiday config bug in Cucumber
For some reason, business_time would occasionally fail when looking up
its configuration to see what the holidays are. This seems to be an
effect of it storing its configuration in Thread.main and the main
thread being a Poltergeist one for Cucumber JS scenarios. I have added
some code to force it to use a class variable instead and that seems to
work.

See https://github.com/18F/micropurchase/issues/1318 for more details on
the bug.