Created CreateAuction class to wrap BuildAuction and SaveAuction
CreateAuction makes it easier to implement and test calls to ChangeState.
In CreateAuction we will make explicit an previously implicit state: 'unpublished'.

unpublished is the default published state for auctions thanks to a default call in a migration.
But we will also ensure that any newly-created auctions also get a State object created reflecting this state.